backend-functions/function_app.py
"""
Azure Functions App for Nutritional Insights
Serverless function endpoints for nutrition data analysis
"""

import azure.functions as func
import logging
import json
import pandas as pd
import os
from io import StringIO

app = func.FunctionApp(http_auth_level=func.AuthLevel.ANONYMOUS)

# Azure Blob Storage connection string (set in Azure portal)
STORAGE_CONNECTION_STRING = os.getenv('AzureWebJobsStorage')
CONTAINER_NAME = 'nutritiondata'
BLOB_NAME = 'All_Diets.csv'

def load_nutrition_data():
    """Load nutrition data from Azure Blob Storage"""
    try:
        from azure.storage.blob import BlobServiceClient
        
        blob_service_client = BlobServiceClient.from_connection_string(STORAGE_CONNECTION_STRING)
        blob_client = blob_service_client.get_blob_client(container=CONTAINER_NAME, blob=BLOB_NAME)
        
        # Download blob content
        blob_data = blob_client.download_blob().readall()
        df = pd.read_csv(StringIO(blob_data.decode('utf-8')))
        
        return df
    except Exception as e:
        logging.error(f"Error loading CSV from blob: {e}")
        return None

def calculate_diet_summary(df):
    """Calculate average macronutrients by diet type"""
    summary = df.groupby('Diet_type').agg({
        'Protein(g)': 'mean',
        'Carbs(g)': 'mean',
        'Fat(g)': 'mean',
        'Recipe_name': 'count'
    }).reset_index()
    
    summary.columns = ['Diet_type', 'Protein', 'Carbs', 'Fat', 'recipes']
    return summary.to_dict('records')

@app.route(route="health", methods=["GET"])
def health_check(req: func.HttpRequest) -> func.HttpResponse:
    """Health check endpoint"""
    logging.info('Health check endpoint called')
    
    return func.HttpResponse(
        json.dumps({
            'status': 'healthy',
            'message': 'Nutritional Insights API is running'
        }),
        mimetype="application/json",
        status_code=200
    )

@app.route(route="nutrition/summary", methods=["GET"])
def get_nutrition_summary(req: func.HttpRequest) -> func.HttpResponse:
    """Get aggregated nutrition summary by diet type"""
    logging.info('Nutrition summary endpoint called')
    
    try:
        df = load_nutrition_data()
        
        if df is None:
            return func.HttpResponse(
                json.dumps({'error': 'Failed to load data from blob storage'}),
                mimetype="application/json",
                status_code=500
            )
        
        diet_summary = calculate_diet_summary(df)
        
        response_data = {
            'status': 'success',
            'total_records': len(df),
            'diet_types': len(diet_summary),
            'data': diet_summary
        }
        
        return func.HttpResponse(
            json.dumps(response_data),
            mimetype="application/json",
            status_code=200
        )
    except Exception as e:
        logging.error(f"Error in nutrition summary: {e}")
        return func.HttpResponse(
            json.dumps({'error': str(e)}),
            mimetype="application/json",
            status_code=500
        )

@app.route(route="recipes/top-protein", methods=["GET"])
def get_top_protein(req: func.HttpRequest) -> func.HttpResponse:
    """Get top protein-rich recipes"""
    logging.info('Top protein endpoint called')
    
    try:
        limit = int(req.params.get('limit', 5))
        df = load_nutrition_data()
        
        if df is None:
            return func.HttpResponse(
                json.dumps({'error': 'Failed to load data'}),
                mimetype="application/json",
                status_code=500
            )
        
        cols = ['Recipe_name', 'Protein(g)', 'Carbs(g)']
        top_recipes = df.nlargest(limit, 'Protein(g)')[cols]
        
        result = []
        for _, row in top_recipes.iterrows():
            result.append({
                'recipe': row['Recipe_name'],
                'protein': round(row['Protein(g)'], 1),
                'carbs': round(row['Carbs(g)'], 1)
            })
        
        return func.HttpResponse(
            json.dumps({
                'status': 'success',
                'count': len(result),
                'data': result
            }),
            mimetype="application/json",
            status_code=200
        )
    except Exception as e:
        logging.error(f"Error in top protein: {e}")
        return func.HttpResponse(
            json.dumps({'error': str(e)}),
            mimetype="application/json",
            status_code=500
        )

@app.route(route="recipes", methods=["GET"])
def get_recipes(req: func.HttpRequest) -> func.HttpResponse:
    """Get recipes with optional filtering"""
    logging.info('Recipes endpoint called')
    
    try:
        df = load_nutrition_data()
        
        if df is None:
            return func.HttpResponse(
                json.dumps({'error': 'Failed to load data'}),
                mimetype="application/json",
                status_code=500
            )
        
        # Optional diet type filter
        diet_type = req.params.get('diet_type')
        if diet_type:
            df = df[df['Diet_type'].str.lower() == diet_type.lower()]
        
        limit = int(req.params.get('limit', 100))
        
        # Calculate statistics
        stats = {
            'total_recipes': len(df),
            'avg_protein': round(df['Protein(g)'].mean(), 2),
            'avg_carbs': round(df['Carbs(g)'].mean(), 2),
            'avg_fat': round(df['Fat(g)'].mean(), 2)
        }
        
        # Get recipe data
        recipes_df = df.head(limit)
        recipes = []
        for _, row in recipes_df.iterrows():
            recipes.append({
                'Recipe_name': row.get('Recipe_name', 'Unknown'),
                'Diet_type': row.get('Diet_type', 'N/A'),
                'Cuisine_type': row.get('Cuisine_type', 'american'),
                'Protein': round(row.get('Protein(g)', 0), 1),
                'Carbs': round(row.get('Carbs(g)', 0), 1),
                'Fat': round(row.get('Fat(g)', 0), 1)
            })
        
        return func.HttpResponse(
            json.dumps({
                'status': 'success',
                'count': len(recipes),
                'recipes': recipes,
                'statistics': stats
            }),
            mimetype="application/json",
            status_code=200
        )
    except Exception as e:
        logging.error(f"Error in recipes: {e}")
        return func.HttpResponse(
            json.dumps({'error': str(e)}),
            mimetype="application/json",
            status_code=500
        )

@app.route(route="clusters", methods=["GET"])
def get_clusters(req: func.HttpRequest) -> func.HttpResponse:
    """Get diet type clusters"""
    logging.info('Clusters endpoint called')
    
    try:
        df = load_nutrition_data()
        
        if df is None:
            return func.HttpResponse(
                json.dumps({'error': 'Failed to load data'}),
                mimetype="application/json",
                status_code=500
            )
        
        summary = calculate_diet_summary(df)
        
        # Simple clustering logic
        high_protein = [d['Diet_type'] for d in summary if d['Protein'] > 90]
        high_carb = [d['Diet_type'] for d in summary if d['Carbs'] > 200]
        balanced = [
            d['Diet_type'] for d in summary
            if d['Diet_type'] not in high_protein
            and d['Diet_type'] not in high_carb
        ]
        
        return func.HttpResponse(
            json.dumps({
                'status': 'success',
                'clusters_identified': 3,
                'high_protein_cluster': high_protein,
                'high_carb_cluster': high_carb,
                'balanced_cluster': balanced
            }),
            mimetype="application/json",
            status_code=200
        )
    except Exception as e:
        logging.error(f"Error in clusters: {e}")
        return func.HttpResponse(
            json.dumps({'error': str(e)}),
            mimetype="application/json",
            status_code=500
        )

@app.route(route="security/status", methods=["GET"])
def get_security_status(req: func.HttpRequest) -> func.HttpResponse:
    """Get security status"""
    return func.HttpResponse(
        json.dumps({
            'status': 'success',
            'security': {
                'encryption': 'Enabled',
                'access_control': 'Secure',
                'compliance': 'GDPR Compliant'
            }
        }),
        mimetype="application/json",
        status_code=200
    )

@app.route(route="auth/2fa/verify", methods=["POST"])
def verify_2fa(req: func.HttpRequest) -> func.HttpResponse:
    """Verify 2FA code"""
    try:
        req_body = req.get_json()
        code = req_body.get('code', '')
        
        if code and len(code) == 6 and code.isdigit():
            return func.HttpResponse(
                json.dumps({
                    'status': 'success',
                    'message': '2FA verification successful',
                    'verified': True
                }),
                mimetype="application/json",
                status_code=200
            )
        else:
            return func.HttpResponse(
                json.dumps({
                    'status': 'error',
                    'message': 'Invalid 2FA code',
                    'verified': False
                }),
                mimetype="application/json",
                status_code=400
            )
    except Exception as e:
        return func.HttpResponse(
            json.dumps({'error': str(e)}),
            mimetype="application/json",
            status_code=500
        )

@app.route(route="cloud/cleanup", methods=["POST"])
def cleanup_resources(req: func.HttpRequest) -> func.HttpResponse:
    """Simulate cloud resource cleanup"""
    import random
    
    storage_freed = round(random.uniform(1.5, 3.5), 1)
    instances_removed = random.randint(2, 5)
    
    return func.HttpResponse(
        json.dumps({
            'status': 'success',
            'message': 'Cloud resource cleanup completed',
            'resources_freed': {
                'storage_gb': storage_freed,
                'compute_instances': instances_removed,
                'actions': [
                    'Removed unused storage containers',
                    'Deleted temporary compute instances',
                    'Cleaned up old log files',
                    'Released unused IP addresses'
                ]
            }
        }),
        mimetype="application/json",
        status_code=200
    )